requirements.txt
GitPython==3.1.31
GitPython==3.1.37
groundingdino.egg==info
gurobipy==10.0.3
numpy==1.23.0
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.7.0.72
Pillow==9.3.0
Pillow==10.0.1
progressbar33==2.4
PuLP==2.7.0
pycocotools==2.0.6
scipy==1.11.3
segment_anything.egg==info
supervision==0.6.0
timm==0.4.12
torch==1.13.0
torchvision==0.14.0
tqdm==4.66.1
ultralytics==8.0.90
